requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
Automat==20.2.0
beautifulsoup4==4.9.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.1
cssselect==1.1.0
discord.py==1.3.3
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
joblib==0.14.1
lxml==4.5.0
multidict==4.7.5
numpy==1.18.3
parsel==1.5.2
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pymongo==3.10.1
pyOpenSSL==19.1.0
queuelib==1.5.0
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
Scrapy==2.0.1
service-identity==18.1.0
six==1.14.0
sklearn==0.0
soupsieve==2.0
Twisted==20.3.0
urllib3==1.25.9
w3lib==1.21.0
websockets==8.1
yarl==1.4.2
zope.interface==5.1.0
